feat: display photo in storefront details
