+++ src/app.pcss
@@ -134,8 +134,8 @@ html, body { @apply h-full overflow-hidden; }
     }
 
     .avatar-container {
-        width: 640px; /* Adjust the width and height as needed */
-        height: 360px;
+        width: 480px; /* Adjust the width and height as needed */
+        height: 240px;
         position: relative;
         overflow: hidden;
         border: 4px solid #ccc;
