+++ src/routes/home_page/components/modal.svelte
@@ -24,12 +24,14 @@
         margin: 10% auto; /* Centering the modal vertically */
         min-width: 40%;
         max-width: 90%;
+        max-height:85%;
         width: auto; /* Adjusting width for responsiveness */
         text-align: center;
         background: white;
         position: fixed;
-        top: 250px; /* Centering the modal vertically */
+        top: 280px; /* Centering the modal vertically */
         left: 50%; /* Centering the modal horizontally */
         transform: translate(-50%, -50%); /* Centering the modal */
+        overflow-y: scroll;
     }
 </style>
